+++ core/src/main/scala/net/ruippeixotog/scalascraper/scraper/ContentParsers.scala
@@ -2,8 +2,8 @@ package net.ruippeixotog.scalascraper.scraper
 
 import scala.util.matching.Regex
 
-import org.joda.time.DateTime
-import org.joda.time.format.{ DateTimeFormat, DateTimeFormatterBuilder }
+import org.joda.time.{ DateTime, DateTimeZone, LocalDate }
+import org.joda.time.format.{ DateTimeFormat, DateTimeFormatter, DateTimeFormatterBuilder }
 
 object ContentParsers {
   def asIs[C] = identity[C] _
@@ -11,10 +11,30 @@ object ContentParsers {
   val asInt: String => Int = _.toInt
   val asDouble: String => Double = _.toDouble
 
-  def asDate(dateFormats: String*): String => DateTime = {
-    val dateParsers = dateFormats.map(DateTimeFormat.forPattern(_).getParser)
-    val formatter = new DateTimeFormatterBuilder().append(null, dateParsers.toArray).toFormatter
-    formatter.parseDateTime
+  @deprecated("Use either asDateTime or asLocalDate", "2.0.0")
+  def asDate(formats: String*) = new AsDateTime(formats)
+
+  def asDateTime(formats: String*) = new AsDateTime(formats)
+  def asLocalDate(formats: String*) = new AsLocalDate(formats)
+
+  def regexMatch(regex: String): RegexMatch = new RegexMatch(regex.r)
+  def regexMatch(regex: Regex): RegexMatch = new RegexMatch(regex)
+  def regexMatches(regex: String): RegexMatches = new RegexMatches(regex.r)
+  def regexMatches(regex: Regex): RegexMatches = new RegexMatches(regex)
+
+  def seq[C, A](parser: C => A): TraversableOnce[C] => TraversableOnce[A] = _.map(parser)
+
+  class AsJodaTime[A](formats: Seq[String], parse: (DateTimeFormatter, String) => A) extends (String => A) {
+    protected[this] lazy val dateParsers = formats.map(DateTimeFormat.forPattern(_).getParser)
+    protected[this] lazy val formatter = new DateTimeFormatterBuilder().append(null, dateParsers.toArray).toFormatter
+
+    def apply(content: String) = parse(formatter, content)
+  }
+
+  class AsLocalDate private[ContentParsers] (formats: Seq[String]) extends AsJodaTime(formats, _.parseLocalDate(_))
+
+  class AsDateTime private[ContentParsers] (formats: Seq[String]) extends AsJodaTime(formats, _.parseDateTime(_)) {
+    def withZone(tz: DateTimeZone): String => DateTime = formatter.withZone(tz).parseDateTime(_).withZone(tz)
   }
 
   class RegexMatch private[ContentParsers] (regex: Regex) extends (String => String) {
@@ -28,11 +48,4 @@ object ContentParsers {
     def captured: String => Iterator[String] = regex.findAllMatchIn(_).map(_.subgroups.head)
     def allCaptured: String => Iterator[List[String]] = regex.findAllMatchIn(_).map(_.subgroups)
   }
-
-  def regexMatch(regex: String): RegexMatch = new RegexMatch(regex.r)
-  def regexMatch(regex: Regex): RegexMatch = new RegexMatch(regex)
-  def regexMatches(regex: String): RegexMatches = new RegexMatches(regex.r)
-  def regexMatches(regex: Regex): RegexMatches = new RegexMatches(regex)
-
-  def seq[C, A](parser: C => A): TraversableOnce[C] => TraversableOnce[A] = _.map(parser)
 }